+++ src/main/java/com/himalayanbus/persistence/repository/IRoleRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface IRoleRepository extends JpaRepository<Role, Integer> {
+public interface IRoleRepository extends JpaRepository<Role, Long> {
 
     Role findByRole(UserRole userRole);
