+++ src/main/java/com/himalayanbus/persistence/repository/IRouteRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IRouteRepository extends JpaRepository<Route, Integer> {
+public interface IRouteRepository extends JpaRepository<Route, Long> {
 
 
     Route findByRouteFromAndRouteTo(String routeFrom, String routeTo);
